engine_spx2html: move initialization into its own module
